refactor again + add reset button
